Passing tests after portfolio refactor
