+++ __tests__/usercard.test.js
@@ -4,7 +4,7 @@
 
 import React from 'react';
 import {render, screen, cleanup} from '@testing-library/react';
-import Usercard from '../client/src/components/Usercard.jsx';
+import Usercard from '../client/src/components/Portfolio/Usercard.jsx';
 
 beforeEach(() => {
   const user = {
@@ -14,31 +14,8 @@ beforeEach(() => {
     email: '',
     cashBalance: 200000,
     rank: 2,
-    userPortfolio: [
-      {
-        stockName: 'Amazon.com, Inc.',
-        stockSymbol: 'AMZN',
-        amount: 350000
-      },
-      {
-        stockName: 'Telsa, Inc.',
-        stockSymbol: 'TSLA',
-        amount: 300000
-      },
-      {
-        stockName: 'Apple',
-        stockSymbol: 'AAPL',
-        amount: 200000
-      },
-      {
-        stockName: 'StockDucks, Inc.',
-        stockSymbol: 'STKD',
-        amount: 200000
-      }
-    ],
-    friends: [
-      // username, username
-    ]
+    userPortfolio: [],
+    portfolioValue: 1050000
   };
   render(<Usercard user={user}/>);
 });
@@ -49,7 +26,7 @@ afterEach(() => {
 
 test('Usercard renders username and rank', () => {
   expect(screen.getByText('RubberDuck')).toBeInTheDocument();
-  expect(screen.getByText('Rank: 2')).toBeInTheDocument();
+  // expect(screen.getByText('Rank: 2')).toBeInTheDocument();
 });
 
 test('Usercard renders cash balance and stock value', () => {
