--- client/src/components/Portfolio/margoNotes.jsx
@@ -1,10 +1,0 @@
-first_name: 'mark', last_name: 'zuckerberg', username: 'the_zuck', email: 'mark_zuckerberg@example.com'
-cashBalance: "1000000"
-email: "mark_zuckerberg@example.com"
-first_name: "mark"
-friends: [{…}]
-last_name: "zuckerberg"
-portfolioValue: NaN
-userPortfolio: [
-{id: 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a23', user_id: 'a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a13', ticker_symbol: 'fb', exchange: 'nasdaq', amount: 1000000} ]
-username: "the_zuck"
